requirements.txt
Django==4.2.2
djangorestframework==3.14.0
requests==2.31.0
django-environ==0.10.0
spotipy==2.23.0
```
